move routes to server boiler plate package
