--- packages/lesmis-server/src/routes/LoginRoute.ts
@@ -1,30 +1,0 @@
-/**
- * Tupaia
- * Copyright (c) 2017 - 2020 Beyond Essential Systems Pty Ltd
- */
-
-import { PermissionsError } from '@tupaia/utils';
-import { AuthConnection, UnauthenticatedRoute } from '@tupaia/server-boilerplate';
-import { Credentials } from '../types';
-import { LESMIS_PERMISSION_GROUP } from '../constants';
-
-export class LoginRoute extends UnauthenticatedRoute {
-  async buildResponse() {
-    const credentials: Credentials = this.req.body;
-
-    const authConnection = new AuthConnection();
-    const response = await authConnection.login(credentials);
-
-    const { id, email, accessPolicy } = await this.sessionModel.createSession(response);
-
-    const authorized = accessPolicy.allowsAnywhere(LESMIS_PERMISSION_GROUP);
-    if (!authorized) {
-      throw new PermissionsError('User not authorized for Lesmis');
-    }
-
-    // set sessionId cookie
-    this.req.sessionCookie = { id, email };
-
-    return { user: response.user };
-  }
-}
