+++ packages/server-boilerplate/README.md
@@ -1,3 +1,7 @@
 # @tupaia/server-boilerplate
 
 Generic boilerplate for back end orchestration servers.
+
+- how to use createApp?
+- need to attach session model?
+
