+++ packages/server-boilerplate/src/connections/index.ts
@@ -1,0 +1,8 @@
+/*
+ * Tupaia
+ * Copyright (c) 2017 - 2020 Beyond Essential Systems Pty Ltd
+ *
+ */
+
+export { AuthConnection } from './AuthConnection';
+export { ApiConnection } from './ApiConnection';
