+++ packages/server-boilerplate/src/models/index.ts
@@ -1,6 +1,7 @@
-/**
+/*
  * Tupaia
  * Copyright (c) 2017 - 2020 Beyond Essential Systems Pty Ltd
+ *
  */
 
-export { UserRoute } from './UserRoute';
+export { SessionType, SessionModel } from './Session';
