+++ packages/server-boilerplate/src/routes/LogoutRoute.ts
@@ -1,9 +1,10 @@
-/**
+/*
  * Tupaia
  * Copyright (c) 2017 - 2020 Beyond Essential Systems Pty Ltd
+ *
  */
 
-import { UnauthenticatedRoute } from '@tupaia/server-boilerplate';
+import { UnauthenticatedRoute } from './UnauthenticatedRoute';
 
 export class LogoutRoute extends UnauthenticatedRoute {
   async buildResponse() {
