+++ packages/server-boilerplate/src/routes/Route.ts
@@ -6,8 +6,8 @@
 import { NextFunction, Response } from 'express';
 import { respond, UnauthenticatedError } from '@tupaia/utils';
 
-import { TupaiaRequest, TupaiaResponseBody, SessionCookie } from './types';
-import { SessionModel, SessionType } from './Session';
+import { TupaiaRequest, TupaiaResponseBody, SessionCookie } from '../types';
+import { SessionModel, SessionType } from '../models/Session';
 
 export abstract class Route {
   req: TupaiaRequest;
