+++ packages/server-boilerplate/src/routes/index.ts
@@ -1,0 +1,11 @@
+/*
+ * Tupaia
+ * Copyright (c) 2017 - 2020 Beyond Essential Systems Pty Ltd
+ *
+ */
+
+export { LoginRoute } from './LoginRoute';
+export { LogoutRoute } from './LogoutRoute';
+export { Route } from './Route';
+export { TestRoute } from './TestRoute';
+export { UnauthenticatedRoute } from './UnauthenticatedRoute';
